Move satisfies_hint to chia._tests.wallet.test_singleton_lifecycle_fast
