--- chia/util/misc.py
@@ -1,38 +1,0 @@
-from __future__ import annotations
-
-from typing import Type, TypeVar, get_args, get_origin
-
-T = TypeVar("T")
-
-
-def satisfies_hint(obj: T, type_hint: Type[T]) -> bool:
-    """
-    Check if an object satisfies a type hint.
-    This is a simplified version of `isinstance` that also handles generic types.
-    """
-    # Start from the initial type hint
-    object_hint_pairs = [(obj, type_hint)]
-    while len(object_hint_pairs) > 0:
-        obj, type_hint = object_hint_pairs.pop()
-        origin = get_origin(type_hint)
-        args = get_args(type_hint)
-        if origin:
-            # Handle generic types
-            if not isinstance(obj, origin):
-                return False
-            if len(args) > 0:
-                # Tuple[T, ...] gets handled just like List[T]
-                if origin is list or (origin is tuple and args[-1] is Ellipsis):
-                    object_hint_pairs.extend((item, args[0]) for item in obj)
-                elif origin is tuple:
-                    object_hint_pairs.extend((item, arg) for item, arg in zip(obj, args))
-                elif origin is dict:
-                    object_hint_pairs.extend((k, args[0]) for k in obj.keys())
-                    object_hint_pairs.extend((v, args[1]) for v in obj.values())
-                else:
-                    raise NotImplementedError(f"Type {origin} is not yet supported")
-        else:
-            # Handle concrete types
-            if type(obj) is not type_hint:
-                return False
-    return True
